apps: add support for MASQUE tunnelled HTTP/3
The major change here is adding support for a QUIC and HTTP/3
connection to contain an "inner" QUIC and HTTP/3 connection.

To make things easier to reason about, the apps library code has
reorganised for better modularity. To support the new functionality
strucutes and methods have been tweaked a bit, and some new methods
added.