apps: properly reject bad CONNECT request
There are a few cases where a CONNECT could be bad, such as being
malformed (as defined in the specification) or failing some other
application condition.

Previously, when quiche-server was run in MASQUE proxy mode, it
validated requests but if there was any error it exited early
without actually sending a response.

This change ensures a response is sent and adds checking in the
client to make sure that the connection is closed if all
CONNECT streams are terminated by the server.