Add MASQUE server relay mode.
Previously, running `quiche-server --masque` would change it into
a forward proxy mode, where client CONNECT/CONNECT-UDP  requests
open TCP/UDP directly to the indicated target.

This change allows the quiche-server to also act as a MASQUE
relay. If the `UDP_PROXY` environment variable is set, when a client
makes a CONNECT-UDP request, we make an indpendent MASQUE connection
to the server's own proxy. Then we shuffle requests and DATAGRAMS
between the two connections in a 1:1 manner.

Since the `quiche-server` is now also a client in this mode, it
has many of the same concerns as `quiche-client`. This change adds
a `--no-verify` option to `quiche-server` that lets it skip TLS
problems. Most other parameters are hard coded or inherited from
the server arguments.

This mode added a bunch of conditional logic to already large code
sections. So the code has been refactored to help readabilit. Hence
the changes look greater that they might actually be.

Testing, this capability is not immediately obvious. The following
example shows one possible deployment.

1. Run quiche-server as the target origin:

  quiche-server --listen 127.0.0.1:9000

2. Run quiche-server as a MASQUE server:

  quiche-server --masque --listen 127.0.0.1:8000

3. Run quiche-server as a MASQUE relay:

  UDP_PROXY=https://127.0.0.1:8000 quiche-server --masque --no-verify --listen 127.0.0.1:7000

4. Run quiche-client as a MASQUE client via the relay:

  UDP_PROXY=https://127.0.0.1:7000 quiche-client --no-verify https://127.0.0.1:9000